pyproject.toml
[project]
name = "peakframer"
dynamic = ["version"]
description = "Extract n-maximally diverse frames from video using visual embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "open-clip-torch>=3.3.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

[project.scripts]
peakframer = "peakframer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "peakframer/__init__.py"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
